task-4-scraping-newspapers/Final_Scraper/requirements.txt
textblob
streamlit
pandas 
nltk
newspaper3k
pygooglenews
pillow
requests
regex
beautifulsoup4
